+++ src/chaparralapi/__init__.py
@@ -4,4 +4,4 @@ Chaparral API
 
 from .client import Client
 
-__version__ = "0.1.1"
+__version__ = "0.2.0"