+++ src/chaparralapi/models/__init__.py
@@ -6,6 +6,6 @@ from .project import Project
 from .project_file import ProjectFile
 from .search_config import (SearchConfig, QuantConfig, TmtConfig, TolConfig, LinearModelConfig, DesignConfig,
                             DatabaseConfig, EnzymeConfig, LfqConfig, StaticModsConfig, VariableModsConfig)
-from .search_result import SearchResult
+from .search_result import SearchResult, PeptideResult, FragmentData, ScanData
 from .search_result_download import SearchResultDownload
 from .search_result_qc import QcScore, QcId, QcPrecursor