+++ src/chaparralapi/routes/project_file.py
@@ -1,4 +1,4 @@
-from typing import Optional, List, Dict, Any
+from typing import Optional, List, Dict, Any, IO
 
 import requests
 from requests import Timeout
@@ -13,19 +13,18 @@ def get_project_files(token: str, project_id: str, base_url: str = DEFAULT_BASE_
     return get(token, url, timeout)
 
 
-def upload_project_file(token: str, project_id: str, file_bytes: bytes, filename: str,
+def upload_project_file(token: str, project_id: str, file: IO, filename: str,
                         base_url: str = DEFAULT_BASE_URL,
                         timeout: Optional[int] = None) -> None:
     url = f"{base_url}/{DEFAULT_PROJECTS_ENDPOINT}/{project_id}/{DEFAULT_PROJECT_FILES_ENDPOINT}"
     if filename.endswith('.raw'):
         files = {
-            'file': (filename, file_bytes, 'application/octet-stream')
+            'file': (filename, file, 'image/RAW')
         }
     else:
         files = {
-            'file': (filename, file_bytes, 'image/RAW')
+            'file': (filename, file, 'application/octet-stream')
         }
-
     try:
         response = requests.put(url, headers={'Authorization': f'Bearer {token}'}, files=files, timeout=timeout)
         response.raise_for_status()