+++ CHANGES.md
@@ -3,6 +3,10 @@
 ## TODO
 - Add more tests
 
+## [0.2.0]
+- added endpoints for get_spectra, get_annotation, get_peptides, test_connection
+- created results module which sets up ProteinGroup, Protein, Peptide, and Psm Iterables, which all return generators of the other iterables.
+
 ## [0.1.0]
 - Added remaining endpoints
 - updated to models/routes structure